requirements.txt
paho-mqtt
rpi-lgpio
pyserial